Refactor: Phase 2 - Extract remaining operations to modular dispatcher
Extract all remaining inline operation handlers from core dispatcher to dedicated
operation modules, completing the modularization initiative started in Phase 1.

## Key Changes

**New Operation Modules:**
- operations/details.ts - Details retrieval operations
- operations/notes.ts - Notes read/create operations
- operations/lists.ts - List management operations
- operations/crud.ts - Create, update, delete operations

**Core Dispatcher Improvements:**
- Reduced core.ts from 500+ to 180 lines
- Removed all inline operation implementations
- Clean modular imports and routing only
- Maintained full backward compatibility

**Architecture Benefits:**
- Modular: Each operation type in focused module
- Maintainable: Easier debugging and testing
- Extensible: Clear pattern for new operations
- Type-Safe: Full TypeScript support preserved

## Technical Details

- Extracted 4 operation handlers with ~300+ lines of code
- Follows Phase 1 patterns for consistency
- Comprehensive parameter validation preserved
- Error handling and response formatting maintained
- No breaking changes to existing tool configurations

Builds on PR #254 (Phase 1) and addresses dispatcher complexity
while preparing foundation for future operation modules.

Resolves #257
