+++ src/handlers/tools/dispatcher/core.ts
@@ -2,20 +2,10 @@
  * Core dispatcher module - main tool execution dispatcher with modular operation handlers
  */
 import { CallToolRequest } from '@modelcontextprotocol/sdk/types.js';
-import { createErrorResult } from '../../../utils/error-handler.js';
-import { parseResourceUri } from '../../../utils/uri-parser.js';
 import { ResourceType } from '../../../types/attio.js';
-import { safeJsonStringify } from '../../../utils/json-serializer.js';
-
-// Import utilities
-import { logToolRequest, logToolError } from './logging.js';
-import { validateAttributes, validateResourceId } from './validation.js';
-import { formatSuccessResponse } from './formatting.js';
 
 // Import tool configurations
 import { findToolConfig } from '../registry.js';
-import { formatResponse } from '../formatters.js';
-import { hasResponseData } from '../error-types.js';
 
 // Import operation handlers
 import {
@@ -24,6 +14,18 @@ import {
   handleSearchByPhone,
   handleSmartSearch
 } from './operations/search.js';
+import { handleDetailsOperation } from './operations/details.js';
+import { 
+  handleNotesOperation, 
+  handleCreateNoteOperation 
+} from './operations/notes.js';
+import { handleGetListsOperation } from './operations/lists.js';
+import {
+  handleCreateOperation,
+  handleUpdateOperation,
+  handleUpdateAttributeOperation,
+  handleDeleteOperation
+} from './operations/crud.js';
 
 // Import tool type definitions
 import {
@@ -168,427 +170,3 @@ export async function executeToolRequest(request: CallToolRequest) {
   }
 }
 
-/**
- * Handle details operation
- */
-async function handleDetailsOperation(
-  request: CallToolRequest,
-  toolConfig: DetailsToolConfig,
-  resourceType: ResourceType
-) {
-  let id: string;
-  let uri: string;
-
-  // Check which parameter is provided
-  const directId =
-    resourceType === ResourceType.COMPANIES
-      ? (request.params.arguments?.companyId as string)
-      : (request.params.arguments?.personId as string);
-
-  uri = request.params.arguments?.uri as string;
-
-  // Use either direct ID or URI, with priority to URI if both are provided
-  if (uri) {
-    try {
-      const [uriType, uriId] = parseResourceUri(uri);
-      if (uriType !== resourceType) {
-        throw new Error(
-          `URI type mismatch: Expected ${resourceType}, got ${uriType}`
-        );
-      }
-      id = uriId;
-    } catch (error) {
-      return createErrorResult(
-        error instanceof Error ? error : new Error('Invalid URI format'),
-        uri,
-        'GET',
-        { status: 400, message: 'Invalid URI format' }
-      );
-    }
-  } else if (directId) {
-    id = directId;
-    // For logging purposes
-    uri = `attio://${resourceType}/${directId}`;
-  } else {
-    return createErrorResult(
-      new Error('Either companyId/personId or uri parameter is required'),
-      `/${resourceType}`,
-      'GET',
-      { status: 400, message: 'Missing required parameter' }
-    );
-  }
-
-  try {
-    const record = await toolConfig.handler(id);
-    const formattedResult = toolConfig.formatResult!(record);
-
-    return formatResponse(formattedResult);
-  } catch (error) {
-    return createErrorResult(
-      error instanceof Error ? error : new Error('Unknown error'),
-      uri,
-      'GET',
-      hasResponseData(error) ? error.response.data : {}
-    );
-  }
-}
-
-/**
- * Handle notes operation
- */
-async function handleNotesOperation(
-  request: CallToolRequest,
-  toolConfig: NotesToolConfig,
-  resourceType: ResourceType
-) {
-  const directId =
-    resourceType === ResourceType.COMPANIES
-      ? (request.params.arguments?.companyId as string)
-      : (request.params.arguments?.personId as string);
-  const uri = request.params.arguments?.uri as string;
-
-  if (!directId && !uri) {
-    const idParamName =
-      resourceType === ResourceType.COMPANIES ? 'companyId' : 'personId';
-    return createErrorResult(
-      new Error(`Either ${idParamName} or uri parameter is required`),
-      `/${resourceType}/notes`,
-      'GET',
-      { status: 400, message: 'Missing required parameter' }
-    );
-  }
-
-  let notesTargetId = directId;
-  let notesResourceType = resourceType;
-
-  try {
-    if (uri) {
-      try {
-        const [uriType, uriId] = parseResourceUri(uri);
-        notesResourceType = uriType as ResourceType;
-        notesTargetId = uriId;
-      } catch (error) {
-        return createErrorResult(
-          error instanceof Error ? error : new Error('Invalid URI format'),
-          uri,
-          'GET',
-          { status: 400, message: 'Invalid URI format' }
-        );
-      }
-    }
-
-    const limit = request.params.arguments?.limit as number;
-    const offset = request.params.arguments?.offset as number;
-
-    const notes = await toolConfig.handler(notesTargetId, limit, offset);
-    const formattedResult = toolConfig.formatResult!(notes);
-
-    return formatResponse(formattedResult);
-  } catch (error) {
-    return createErrorResult(
-      error instanceof Error ? error : new Error('Unknown error'),
-      uri || `/${resourceType}/${notesTargetId}/notes`,
-      'GET',
-      hasResponseData(error) ? error.response.data : {}
-    );
-  }
-}
-
-/**
- * Handle createNote operation
- */
-async function handleCreateNoteOperation(
-  request: CallToolRequest,
-  toolConfig: CreateNoteToolConfig,
-  resourceType: ResourceType
-) {
-  const directId =
-    resourceType === ResourceType.COMPANIES
-      ? (request.params.arguments?.companyId as string)
-      : (request.params.arguments?.personId as string);
-  const uri = request.params.arguments?.uri as string;
-
-  /**
-   * Parameter Mapping Strategy for Note Creation
-   *
-   * This function supports multiple parameter names for backward compatibility
-   * and to accommodate different API clients:
-   *
-   * - title: Primary parameter name (preferred)
-   * - noteTitle: Legacy/alternative parameter name for title
-   * - content: Primary parameter name (preferred)
-   * - noteText: Legacy/alternative parameter name for content
-   *
-   * The fallback pattern (primary || legacy) ensures compatibility while
-   * encouraging use of the standardized parameter names.
-   */
-  const title = (request.params.arguments?.title ||
-    request.params.arguments?.noteTitle) as string;
-  const content = (request.params.arguments?.content ||
-    request.params.arguments?.noteText) as string;
-
-  if (!title || !content) {
-    return createErrorResult(
-      new Error('Both title and content are required'),
-      `/${resourceType}/notes`,
-      'POST',
-      { status: 400, message: 'Missing required parameters' }
-    );
-  }
-
-  if (!directId && !uri) {
-    const idParamName =
-      resourceType === ResourceType.COMPANIES ? 'companyId' : 'personId';
-    return createErrorResult(
-      new Error(`Either ${idParamName} or uri parameter is required`),
-      `/${resourceType}/notes`,
-      'POST',
-      { status: 400, message: 'Missing required parameter' }
-    );
-  }
-
-  let noteTargetId = directId;
-  let noteResourceType = resourceType;
-
-  try {
-    if (uri) {
-      try {
-        const [uriType, uriId] = parseResourceUri(uri);
-        noteResourceType = uriType as ResourceType;
-        noteTargetId = uriId;
-      } catch (error) {
-        return createErrorResult(
-          error instanceof Error ? error : new Error('Invalid URI format'),
-          uri,
-          'POST',
-          { status: 400, message: 'Invalid URI format' }
-        );
-      }
-    }
-
-    const note = await toolConfig.handler(noteTargetId, title, content);
-    const formattedResult = toolConfig.formatResult
-      ? toolConfig.formatResult(note)
-      : `Note added to ${resourceType.slice(0, -1)} ${noteTargetId}: ${note.title || 'Untitled'}`;
-
-    return formatResponse(formattedResult);
-  } catch (error) {
-    return createErrorResult(
-      error instanceof Error ? error : new Error('Unknown error'),
-      uri || `/${resourceType}/${noteTargetId}/notes`,
-      'POST',
-      hasResponseData(error) ? error.response.data : {}
-    );
-  }
-}
-
-/**
- * Handle getLists operation
- */
-async function handleGetListsOperation(
-  request: CallToolRequest,
-  toolConfig: GetListsToolConfig
-) {
-  try {
-    const lists = await toolConfig.handler();
-    const formattedResult = toolConfig.formatResult!(lists);
-
-    return formatResponse(formattedResult);
-  } catch (error) {
-    return createErrorResult(
-      error instanceof Error ? error : new Error('Unknown error'),
-      '/lists',
-      'GET',
-      hasResponseData(error) ? error.response.data : {}
-    );
-  }
-}
-
-/**
- * Handle create operation
- */
-async function handleCreateOperation(
-  request: CallToolRequest,
-  toolConfig: ToolConfig,
-  resourceType: ResourceType
-) {
-  try {
-    const attributes = request.params.arguments?.attributes;
-    
-    if (!attributes) {
-      return createErrorResult(
-        new Error('Attributes parameter is required for create operation'),
-        `/${resourceType}`,
-        'POST',
-        { status: 400, message: 'Missing required parameter: attributes' }
-      );
-    }
-
-    const result = await toolConfig.handler(attributes);
-    const formattedResult = toolConfig.formatResult
-      ? toolConfig.formatResult(result)
-      : `${resourceType.slice(0, -1)} created successfully`;
-
-    return formatResponse(formattedResult);
-  } catch (error) {
-    return createErrorResult(
-      error instanceof Error ? error : new Error('Unknown error'),
-      `/${resourceType}`,
-      'POST',
-      hasResponseData(error) ? error.response.data : {}
-    );
-  }
-}
-
-/**
- * Handle update operation
- */
-async function handleUpdateOperation(
-  request: CallToolRequest,
-  toolConfig: ToolConfig,
-  resourceType: ResourceType
-) {
-  try {
-    const id = resourceType === ResourceType.COMPANIES
-      ? (request.params.arguments?.companyId as string)
-      : (request.params.arguments?.personId as string);
-    
-    const attributes = request.params.arguments?.attributes;
-    
-    if (!id) {
-      const idParamName = resourceType === ResourceType.COMPANIES ? 'companyId' : 'personId';
-      return createErrorResult(
-        new Error(`${idParamName} parameter is required for update operation`),
-        `/${resourceType}`,
-        'PUT',
-        { status: 400, message: `Missing required parameter: ${idParamName}` }
-      );
-    }
-    
-    if (!attributes) {
-      return createErrorResult(
-        new Error('Attributes parameter is required for update operation'),
-        `/${resourceType}/${id}`,
-        'PUT',
-        { status: 400, message: 'Missing required parameter: attributes' }
-      );
-    }
-
-    const result = await toolConfig.handler(id, attributes);
-    const formattedResult = toolConfig.formatResult
-      ? toolConfig.formatResult(result)
-      : `${resourceType.slice(0, -1)} updated successfully`;
-
-    return formatResponse(formattedResult);
-  } catch (error) {
-    return createErrorResult(
-      error instanceof Error ? error : new Error('Unknown error'),
-      `/${resourceType}`,
-      'PUT',
-      hasResponseData(error) ? error.response.data : {}
-    );
-  }
-}
-
-/**
- * Handle updateAttribute operation
- */
-async function handleUpdateAttributeOperation(
-  request: CallToolRequest,
-  toolConfig: ToolConfig,
-  resourceType: ResourceType
-) {
-  try {
-    const id = resourceType === ResourceType.COMPANIES
-      ? (request.params.arguments?.companyId as string)
-      : (request.params.arguments?.personId as string);
-    
-    const attributeName = request.params.arguments?.attributeName as string;
-    const value = request.params.arguments?.value;
-    
-    if (!id) {
-      const idParamName = resourceType === ResourceType.COMPANIES ? 'companyId' : 'personId';
-      return createErrorResult(
-        new Error(`${idParamName} parameter is required for updateAttribute operation`),
-        `/${resourceType}`,
-        'PATCH',
-        { status: 400, message: `Missing required parameter: ${idParamName}` }
-      );
-    }
-    
-    if (!attributeName) {
-      return createErrorResult(
-        new Error('attributeName parameter is required for updateAttribute operation'),
-        `/${resourceType}/${id}`,
-        'PATCH',
-        { status: 400, message: 'Missing required parameter: attributeName' }
-      );
-    }
-    
-    if (value === undefined) {
-      return createErrorResult(
-        new Error('value parameter is required for updateAttribute operation'),
-        `/${resourceType}/${id}`,
-        'PATCH',
-        { status: 400, message: 'Missing required parameter: value' }
-      );
-    }
-
-    const result = await toolConfig.handler(id, attributeName, value);
-    const formattedResult = toolConfig.formatResult
-      ? toolConfig.formatResult(result)
-      : `${resourceType.slice(0, -1)} attribute updated successfully`;
-
-    return formatResponse(formattedResult);
-  } catch (error) {
-    return createErrorResult(
-      error instanceof Error ? error : new Error('Unknown error'),
-      `/${resourceType}`,
-      'PATCH',
-      hasResponseData(error) ? error.response.data : {}
-    );
-  }
-}
-
-/**
- * Handle delete operation
- */
-async function handleDeleteOperation(
-  request: CallToolRequest,
-  toolConfig: ToolConfig,
-  resourceType: ResourceType
-) {
-  try {
-    const id = resourceType === ResourceType.COMPANIES
-      ? (request.params.arguments?.companyId as string)
-      : (request.params.arguments?.personId as string);
-    
-    if (!id) {
-      const idParamName = resourceType === ResourceType.COMPANIES ? 'companyId' : 'personId';
-      return createErrorResult(
-        new Error(`${idParamName} parameter is required for delete operation`),
-        `/${resourceType}`,
-        'DELETE',
-        { status: 400, message: `Missing required parameter: ${idParamName}` }
-      );
-    }
-
-    const result = await toolConfig.handler(id);
-    const formattedResult = toolConfig.formatResult
-      ? toolConfig.formatResult(result)
-      : `${resourceType.slice(0, -1)} deleted successfully`;
-
-    return formatResponse(formattedResult);
-  } catch (error) {
-    const id = resourceType === ResourceType.COMPANIES
-      ? (request.params.arguments?.companyId as string)
-      : (request.params.arguments?.personId as string);
-    
-    return createErrorResult(
-      error instanceof Error ? error : new Error('Unknown error'),
-      `/${resourceType}/${id || 'unknown'}`,
-      'DELETE',
-      hasResponseData(error) ? error.response.data : {}
-    );
-  }
-}
